// src/types/branding/branding-step.enum.ts
export enum BrandingStep {
    BRAND_DISCOVERY = 'BRAND_DISCOVERY',
    VISUAL_INSPIRATION = 'VISUAL_INSPIRATION',
    LOGO_EXPLORATION = 'LOGO_EXPLORATION',
    COLOR_HARMONY = 'COLOR_HARMONY',
    TYPOGRAPHY_SELECTION = 'TYPOGRAPHY_SELECTION',
    IMAGERY_DIRECTION = 'IMAGERY_DIRECTION',
    BRAND_VOICE = 'BRAND_VOICE',
    BRAND_IDENTITY_SUITE = 'BRAND_IDENTITY_SUITE',
  }